Fix calcDamge: stop flooring the attack/defense ratio to an integer
calcDamge computed int(atk/def) before plugging into the damage formula, so
any time the attacking stat was below the defending stat (ratio < 1 — very
common) the ratio floored to 0 and damage collapsed to the +2 base. This
silently capped every damage estimate and distorted move choice, switch
scoring, boosting value, and ability checks alike.

The canonical Gen 5+ formula keeps A/D as a real ratio and only floors the
larger expression. Removed the inner int() on the ratio for both physical
and special branches.

Verified: damage now scales smoothly with the A/D ratio (0.5 -> 54, 1.0 ->
106, 1.75 -> 186 for a 90 BP move) instead of reading ~0 below ratio 1.

+++ libs.py
@@ -21,10 +21,13 @@ def calcDamge(poke1, poke2, move, currentHP):
 
     #find damage before min max multiplier
     damage = 0
+    # Canonical Gen 5+ formula: floor( floor(42 * Power * A/D / 50) + 2 ) * mods.
+    # A/D is a real ratio here; flooring it (the old bug) zeroed out damage
+    # whenever the attacking stat was below the defending stat.
     if move.category == 'Physical':
-        damage = int( ( int( (42 * move.power * int(poke1.stats['atk'] / poke2.stats['defe'])) / 50) + 2) * stab * effective)
+        damage = int( ( int( (42 * move.power * (poke1.stats['atk'] / poke2.stats['defe'])) / 50) + 2) * stab * effective)
     elif move.category == 'special':
-        damage = int( ( int( (42 * move.power * int(poke1.stats['spa'] / poke2.stats['spd'])) / 50) + 2) * stab * effective)
+        damage = int( ( int( (42 * move.power * (poke1.stats['spa'] / poke2.stats['spd'])) / 50) + 2) * stab * effective)
     minDamage = int(damage * MIN_MULT)
     maxDamage = int(damage * MAX_MULT)
